Merge of user's editor rework (TabContainer: Данные/Карта with draggable
MapCanvas/Текстуры/Анимации; CI workflows; clippy fixes; serde float/int
helpers; expanded core content) with this branch's editor v2 features,
re-applied on top of the new tab structure:
- upgraded SCHEMAS (dyn_enum sources, vec2/vec3/color fields, Диалоги,
  Пресет single, flats/glows/ammo/weapon-spawn map categories, maps/*);
- map-file picker in the top bar (data categories AND the visual canvas
  follow it); record search; user://presets; reference validation on save;
- restored the AI texture generation UI (lost in the parallel rework) as a
  dedicated "🎨 ИИ" tab with background thread + teardown on plugin exit;
- missing map sections are created on add.

+++ rust/src/classes.rs
@@ -45,7 +45,7 @@ struct SpecRaw {
 }
 
 #[derive(Deserialize)]
-struct AmmoStartRaw { #[serde(rename = "type")] ty: String, amount: u32 }
+struct AmmoStartRaw { #[serde(rename = "type")] ty: String, amount: f64 }
 
 #[derive(Deserialize)]
 struct ClassRaw {
@@ -83,7 +83,7 @@ impl ClassRaw {
         for a in &self.start_ammo {
             let t = AmmoType::from_id(&a.ty)
                 .ok_or_else(|| format!("class '{}': unknown ammo '{}'", self.id, a.ty))?;
-            ammo.push((t, a.amount));
+            ammo.push((t, a.amount as u32));
         }
         let mut specs = Vec::new();
         for s in self.specs {
